Split up pytransform3d.rotations
